pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "CLI application for expense analysis based on T-Bank PDF statements"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pdfplumber>=0.11.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "openpyxl>=3.1.0",
    "gspread>=6.0.0",
    "google-auth>=2.0.0",
    "aiogram>=3.4.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
expense-tracker = "expense_tracker.cli:app"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
include = ["expense_tracker*"]
